Release v3.7.0.0 - update to meet 3.7 APIs
Updates:
* privacy provider
* new param added to the core_calendar_provide_event_action func
* moves off of using the legacy polyfill

+++ classes/privacy/provider.php
@@ -25,9 +25,11 @@ namespace mod_lightboxgallery\privacy;
 
 use core_privacy\local\metadata\collection;
 use core_privacy\local\request\approved_contextlist;
+use core_privacy\local\request\approved_userlist;
 use core_privacy\local\request\contextlist;
 use core_privacy\local\request\deletion_criteria;
 use core_privacy\local\request\helper;
+use core_privacy\local\request\userlist;
 use core_privacy\local\request\writer;
 
 defined('MOODLE_INTERNAL') || die();
@@ -40,12 +42,10 @@ defined('MOODLE_INTERNAL') || die();
 class provider implements
         // This plugin stores personal data.
         \core_privacy\local\metadata\provider,
+        \core_privacy\local\request\core_userlist_provider,
 
         // This plugin is a core_user_data_provider.
         \core_privacy\local\request\plugin\provider {
-
-    // This trait must be included.
-    use \core_privacy\local\legacy_polyfill;
 
     /**
      * Return the fields which contain personal data.
@@ -53,7 +53,7 @@ class provider implements
      * @param collection $items a reference to the collection to use to store the metadata.
      * @return collection the updated collection of metadata items.
      */
-    public static function _get_metadata(collection $items) {
+    public static function get_metadata(collection $items) : collection {
         $items->add_database_table(
             'lightboxgallery_comments',
             [
@@ -74,7 +74,7 @@ class provider implements
      * @param int $userid the userid.
      * @return contextlist the list of contexts containing user info for the user.
      */
-    public static function _get_contexts_for_userid($userid) {
+    public static function get_contexts_for_userid(int $userid) : contextlist {
         // Fetch all lightboxgallery comments.
         $sql = "SELECT c.id
                 FROM {context} c
@@ -100,7 +100,7 @@ class provider implements
      *
      * @param approved_contextlist $contextlist a list of contexts approved for export.
      */
-    public static function _export_user_data(approved_contextlist $contextlist) {
+    public static function export_user_data(approved_contextlist $contextlist) {
         global $DB;
 
         if (empty($contextlist->count())) {
@@ -174,11 +174,39 @@ class provider implements
     }
 
     /**
+     * Get the list of users who have data within a context.
+     *
+     * @param userlist $userlist The userlist containing the list of users who have data in this context/plugin combination.
+     *
+     */
+    public static function get_users_in_context(userlist $userlist) {
+        $context = $userlist->get_context();
+
+        if (!is_a($context, \context_module::class)) {
+            return;
+        }
+
+        $sql = "SELECT lbgc.userid
+                FROM {context} c
+                JOIN {course_modules} cm ON cm.id = c.instanceid AND c.contextlevel = :contextlevel
+                JOIN {modules} m ON m.id = cm.module AND m.name = :modname
+                JOIN {lightboxgallery} lbg ON lbg.id = cm.instance
+                JOIN {lightboxgallery_comments} lbgc ON lbgc.gallery = lbg.id";
+
+        $params = [
+            'modname'       => 'lightboxgallery',
+            'contextlevel'  => \CONTEXT_MODULE,
+        ];
+
+        $userlist->add_from_sql('userid', $sql, $params);
+    }
+
+    /**
      * Delete all data for all users in the specified context.
      *
      * @param \context $context the context to delete in.
      */
-    public static function _delete_data_for_all_users_in_context(\context $context) {
+    public static function delete_data_for_all_users_in_context(\context $context) {
         global $DB;
 
         if (empty($context)) {
@@ -189,8 +217,41 @@ class provider implements
             return;
         }
 
-        $instanceid = $DB->get_field('course_modules', 'instance', ['id' => $context->instanceid], MUST_EXIST);
+        $instanceid = self::get_lightboxgallery_id_from_context($context);
+        if (empty($instanceid)) {
+            return;
+        }
+
         $DB->delete_records('lightboxgallery_comments', ['gallery' => $instanceid]);
+    }
+
+    /**
+     * Delete multiple users within a single context.
+     *
+     * @param approved_userlist $userlist The approved context and user information to delete information for.
+     */
+    public static function delete_data_for_users(approved_userlist $userlist) {
+        global $DB;
+
+        $context = $userlist->get_context();
+        $instanceid = self::get_lightboxgallery_id_from_context($context);
+        $userids = $userlist->get_userids();
+
+        if (empty($instanceid)) {
+            return;
+        }
+
+        // Prepare the SQL we'll need below.
+        list($insql, $inparams) = $DB->get_in_or_equal($userids, SQL_PARAMS_NAMED);
+        $sql = "gallery = :instanceid AND userid {$insql}";
+        $params = array_merge($inparams, ['instanceid' => $instanceid]);
+
+        $DB->delete_records_select('lightboxgallery_comments', $sql, $params);
+    }
+
+    protected static function get_lightboxgallery_id_from_context(context_module $context) {
+        $cm = get_coursemodule_from_id('lightboxgallery', $context->instanceid);
+        return $cm ? (int) $cm->instance : 0;
     }
 
     /**
@@ -198,7 +259,7 @@ class provider implements
      *
      * @param approved_contextlist $contextlist a list of contexts approved for deletion.
      */
-    public static function _delete_data_for_user(approved_contextlist $contextlist) {
+    public static function delete_data_for_user(approved_contextlist $contextlist) {
         global $DB;
 
         if (empty($contextlist->count())) {
